requirements.txt
diffusers==0.27.2
transformers>=4.36.0
accelerate>=0.25.0
safetensors>=0.4.0
opencv-python-headless>=4.9.0
pillow>=10.0.0
imageio>=2.33.0
torch>=2.1.0
torchvision>=0.16.0
insightface>=0.7.3
onnxruntime-gpu>=1.16.0
runpod>=1.6.0
pydantic>=2.0.0
gradio>=4.19.0
requests>=2.31.0
librosa>=0.10.0
numpy<2.0.0
huggingface_hub>=0.20.0
